- Load tasks from timecamp service

+++ src/Modules/TimeTracker/FocusForge.TimeTracker/Entities/TimeTrackerTask.cs
@@ -8,8 +8,8 @@ namespace FocusForge.TimeTracker.Entities
 {
     public class TimeTrackerTask
     {
-        public int Id { get; set; }
-        public int ExternalId { get; set; }
+        public long Id { get; set; }
+        public long ExternalId { get; set; }
         public string Name { get; set; }
     }
 }
